숫자 버튼에 대한 부분 Digits 컴포넌트로 분리
